Extract ArticleButton
